comentarios

+++ projeto-3/tarefa-a/tarefa-a-10407962.f90
@@ -1,5 +1,5 @@
 program tarefaa
-    implicit real*16 (a-h,o-z)
+    implicit real*8 (a-h,o-z)
 
     dimension h(14)
